Initial work to start using the Perfetto proto format in the Performance tab
This is the work needed to pass the Perfetto proto from the VM service
to the Perfetto trace viewer. If we didn't have to worry about the
Flutter specific functionality and offline mode, we'd basically be done.

However, DevTools makes a lot of assumptions based on the
structure of the Chrome Trace Event format, many of which are deeply
embedded in the logic for handling Flutter frame events. This includes:

- Reaching directly into the backing JSON of TraceEvent to poke at
  some specific fields, something that isn't possible to do with the
  Perfetto TrackEvents.

- Reliance on the CTE format's COMPLETE events, which don't exist in
  the Perfetto format. See:
  https://perfetto.dev/docs/reference/trace-packet-proto#TrackEvent.Type

- Reliance on the CTE's async begin/end events, which don't exist in
  the Perfetto format. I _think_ Perfetto treats sync and async events
  identically, where async events are simply logged to their own track.

These are only the issues I'm aware of; there's probably a lot more that
needs to be dealt with as well. I'm not sure what will need to be done
to get import/export working as well.